move consts in api/admission to consts/admission/consts.go
